My Code

+++ fsrhilmakv2/Models/AccountViewModels.cs
@@ -52,7 +52,7 @@ namespace fsrhilmakv2.Models
 
         public string PictureId { get; set; }
 
-        public int AvgServicesInOneDay { get; set; }
+        public double AvgServicesInOneDay { get; set; }
 
         public int Age { get; set; }
